requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
bs4==0.0.1
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
commonmark==0.9.1
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
entrypoints==0.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
notion==0.0.25
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
